create jpe services 1

+++ repository/src/main/java/nicomed/tms/projectplanner/repository/TitleListRepository.java
@@ -3,5 +3,5 @@ package nicomed.tms.projectplanner.repository;
 import nicomed.tms.projectplanner.entity.TitleList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TitleListRepository extends JpaRepository<TitleList, Long> {
+public interface TitleListRepository extends JpaRepository<TitleList, Integer> {
 }
